tests/workflow-integration/requirements.txt
# Testing dependencies for workflow and integration validation
pytest>=7.0.0
pytest-json-report>=1.5.0
boto3>=1.26.0
moto>=4.2.0
pyyaml>=6.0
requests>=2.28.0

# Additional testing utilities
pytest-cov>=4.0.0
pytest-xdist>=3.0.0
pytest-mock>=3.10.0
pytest-asyncio>=0.21.0

# AWS SDK and utilities
botocore>=1.29.0
awscli>=1.27.0

# Workflow and CI/CD testing
github3.py>=3.2.0
gitpython>=3.1.0

# Code quality and formatting
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0

# Security and validation
bandit>=1.7.0
safety>=2.3.0